Add support for butler
Make it trivial to use makelove to push butler builds to itch.io.

Adds a new butler section with two parameters:

itchapp - the name of the app: username/projectname.
publish_love - whether to publish the .love file (since it's not a
    target).

Remove butler from postbuild examples since we have a better way of
doing that.

Test
====

Created a test project on itch and pushed win32, win64, macos, lovejs,
love builds to it from Win10. Incremented version and pushed a new win32
build. Version numbers show as expected.

Play lovejs build in browser. Play from itch app.

Butler is not invoked when itchapp is not defined.